Refactor PactSwap integration by moving API functions to a new pactswap.ts file
